Moved knexfile to project root for simplicity
