+++ server/models/bookshelf/order.js
@@ -1,6 +1,3 @@
-const path = require('path');
-
-console.log(path.resolve());
 const bookshelf = require('../../config/bookshelf.js');
 
 module.exports = bookshelf.model('Order', {
