+++ server/models/seeds/02_vaccs.js
@@ -1,18 +1,15 @@
 const parser = require('../utils/parser.js');
 const vaccination = require('../utils/vaccination.js');
 const batchInsert = require('../utils/batch_insert.js');
-const getPath = require('../../utils/get_path.js');
 
-const vaccPath = getPath('\\server\\config', '../models/data/vaccinations.source','server/models/data/vaccinations.source');
-
-const environment = process.env.NODE_ENV || 'development';
+const vaccPath = 'server/models/data/vaccinations.source';
 
 exports.seed = async (knex) => {
   try {
     // Getting an array of objects from the source file
     let vaccInsert = await parser(vaccPath, vaccination);
     // The maximum batch size for sqlite is 500, so the insert will be divided into several inserts
-    if (environment == 'development') {
+    if (knex.client.config.client == 'sqlite3') {
       await batchInsert(vaccInsert, knex, 'Vaccination', 500);
     } else {
       await knex('Vaccination').insert(vaccInsert);
